refactor: renomme les fichiers V2 en Editorialisee(s)
