+++ site/source/components/Simulateur/QuestionPublicodes.tsx
@@ -4,7 +4,7 @@ import { useDispatch } from 'react-redux'
 import { ValeurPublicodes } from '@/domaine/engine/PublicodesAdapter'
 import { DottedName } from '@/domaine/publicodes/DottedName'
 import { Situation } from '@/domaine/Situation'
-import { QuestionPublicodes as TypeQuestionPublicodes } from '@/hooks/useQuestionsPublicodesV2'
+import { QuestionPublicodes as TypeQuestionPublicodes } from '@/hooks/useQuestionsPublicodesEditorialisees'
 import { enregistreLaRéponseÀLaQuestion } from '@/store/actions/actions'
 
 import { RuleField } from '../conversation/RuleField'
