+++ site/source/components/Simulateur/ZoneDeSaisie.tsx
@@ -3,7 +3,7 @@ import { styled } from 'styled-components'
 
 import { SmallBody } from '@/design-system'
 import { Situation } from '@/domaine/Situation'
-import { QuestionPublicodes } from '@/hooks/useQuestionsPublicodesV2'
+import { QuestionPublicodes } from '@/hooks/useQuestionsPublicodesEditorialisees'
 
 import { BlocMontants } from './BlocMontants'
 import { BlocSituation } from './BlocSituation'
